+++ src/routes/recipes/+page.svelte
@@ -3,7 +3,6 @@
   import { getAllCoffeeBeans } from "$lib/database/manageCoffeeBeans";
   import { getAllRecipes } from "$lib/database/manageRecipes";
   import { routes } from "$lib/domain/constants/routes";
-  import type { EnhancedRecipe } from "$lib/domain/enhancedEntities/EnhancedRecipe";
   import type { CoffeeBeans } from "$lib/domain/entities/CoffeeBeans";
   import type { Recipe } from "$lib/domain/entities/Recipe";
   import type { RecipesSortOrderEnum } from "$lib/domain/sort/RecipesSortOrderEnum";
@@ -11,6 +10,7 @@
     sortRecipesByTimestampDesc as byTimestampDesc,
     sortRecipesByTimestampDesc
   } from "$lib/domain/sort/sortRecipes";
+  import type { EnhancedRecipe } from "$lib/helperTypes/EnhancedRecipe";
   import RecipeCard from "$lib/UI/domainComponents/cards/RecipeCard.svelte";
   import MyFab from "$lib/UI/domainComponents/FABs/AddRecipeFab.svelte";
   import Loading from "$lib/UI/domainComponents/lists/Loading.svelte";
